Change RSSI datatype to int32_t
* Solves #62

+++ server/include/data_store.hpp
@@ -44,7 +44,7 @@ public:
     bool AssignDeviceToEmployee(const std::string& device_id, const std::string& employee_id);
 
     bool InsertRSSIReadings(const std::string& device_id,
-                            std::vector<std::pair<std::string, double>> macaddr_rssi_datapoints);
+                            std::vector<std::pair<std::string, int32_t>> macaddr_rssi_datapoints);
 
     bool GetPosition(const std::string& device_id, QueryT queryby, Position& pos);
